feat: Added views for add products manually
